video-platform/backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
Pillow==11.0.0
python-decouple==3.8
django-cleanup==9.0.0
drf-yasg==1.21.8
